+++ reductstore/src/storage/block_manager.rs
@@ -42,6 +42,8 @@ pub struct BlockManager {
     path: PathBuf,
     use_counter: UseCounter,
     file_cache: FileCache,
+    bucket: String,
+    entry: String,
     last_block: Option<Block>,
 }
 
@@ -74,10 +76,20 @@ pub fn find_first_block(block_index: &BTreeSet<u64>, start: &u64) -> u64 {
 
 impl BlockManager {
     pub(crate) fn new(path: PathBuf) -> Self {
+        let (bucket, entry) = {
+            let mut parts = path.iter().rev();
+            let entry = parts.next().unwrap().to_str().unwrap().to_string();
+            let bucket = parts.next().unwrap().to_str().unwrap().to_string();
+            (bucket, entry)
+        };
+
         Self {
             path,
             use_counter: UseCounter::new(IO_OPERATION_TIMEOUT),
             file_cache: FileCache::new(FILE_CACHE_MAX_SIZE, FILE_CACHE_TIME_TO_LIVE),
+            bucket,
+            entry,
+
             last_block: None,
         }
     }
@@ -101,7 +113,7 @@ impl BlockManager {
         self.use_counter.increment(ts_to_us(&ts));
 
         let path = self.path_to_data(&ts);
-        let file = self.file_cache.read(&path).await?;
+        let file = self.file_cache.write_or_create(&path).await?;
         let offset = block.records[record_index].begin;
         Ok((file, offset as usize))
     }
@@ -122,8 +134,8 @@ impl BlockManager {
 
         self.save(block).await?;
         debug!(
-            "Finished writing record {} with state {:?}",
-            timestamp, state
+            "Finished writing record {}/{}/{} with state {:?}",
+            self.bucket, self.entry, timestamp, state
         );
 
         Ok(())
@@ -139,7 +151,7 @@ impl BlockManager {
         self.use_counter.increment(ts_to_us(&ts));
 
         let path = self.path_to_data(&ts);
-        let file = self.file_cache.write_or_create(&path).await?;
+        let file = self.file_cache.read(&path).await?;
         let offset = block.records[record_index].begin;
         Ok((file, offset as usize))
     }
@@ -217,19 +229,20 @@ impl ManageBlock for BlockManager {
             }
         }
 
-        let proto_ts = us_to_ts(&block_id);
-        let file = self
-            .file_cache
-            .write_or_create(&self.path_to_desc(&proto_ts))
-            .await?;
+        let path = self.path_to_desc(&us_to_ts(&block_id));
+        let file = self.file_cache.read(&path).await?;
         let mut buf = vec![];
 
         // parse the block descriptor
         let mut lock = file.write().await;
         lock.seek(SeekFrom::Start(0)).await?;
         lock.read_to_end(&mut buf).await?;
+
         let block = Block::decode(Bytes::from(buf)).map_err(|e| {
-            ReductError::internal_server_error(&format!("Failed to decode block descriptor: {}", e))
+            ReductError::internal_server_error(&format!(
+                "Failed to decode block descriptor {:?}: {}",
+                path, e
+            ))
         })?;
 
         Ok(block)
@@ -239,7 +252,10 @@ impl ManageBlock for BlockManager {
         let path = self.path_to_desc(block.begin_time.as_ref().unwrap());
         let mut buf = BytesMut::new();
         block.encode(&mut buf).map_err(|e| {
-            ReductError::internal_server_error(&format!("Failed to encode block descriptor: {}", e))
+            ReductError::internal_server_error(&format!(
+                "Failed to encode block descriptor {:?}: {}",
+                path, e
+            ))
         })?;
 
         // overwrite the file
@@ -248,7 +264,6 @@ impl ManageBlock for BlockManager {
         lock.set_len(buf.len() as u64).await?;
         lock.seek(SeekFrom::Start(0)).await?;
         lock.write_all(&buf).await?;
-        lock.flush().await?;
 
         self.last_block = Some(block);
         Ok(())
@@ -386,7 +401,11 @@ pub async fn spawn_read_task(
                 break;
             }
             if let Err(e) = tx.send(Ok(Bytes::from(buf))).await {
-                debug!("Failed to send record {} chunk: {}", record_ts, e); // for some reason the receiver is closed
+                let bm = block_manager.read().await;
+                debug!(
+                    "Failed to send record {}/{}/{}: {}",
+                    bm.bucket, bm.entry, record_ts, e
+                ); // for some reason the receiver is closed
                 break;
             }
 
@@ -484,7 +503,11 @@ pub async fn spawn_write_task(
         let state = match recv.await {
             Ok(_) => record::State::Finished,
             Err(err) => {
-                error!("Failed to write record {}: {}", record_ts, err);
+                let bm = bm.read().await;
+                error!(
+                    "Failed to write record {}/{}/{}: {}",
+                    bm.bucket, bm.entry, record_ts, err
+                );
                 if err.status == ErrorCode::InternalServerError {
                     record::State::Invalid
                 } else {
@@ -499,7 +522,11 @@ pub async fn spawn_write_task(
             .finish_write_record(block_id, state, record_index)
             .await
         {
-            error!("Failed to finish writing {} record: {}", record_ts, err);
+            let bm = bm.read().await;
+            error!(
+                "Failed to finish writing {}/{}/{} record: {}",
+                bm.bucket, bm.entry, record_ts, err
+            );
         }
     });
     Ok(tx)
@@ -607,7 +634,7 @@ mod tests {
 
         tx.send(Ok(None)).await.unwrap();
         drop(tx);
-        sleep(std::time::Duration::from_millis(10)).await; // wait for thread to finish
+        sleep(Duration::from_millis(10)).await; // wait for thread to finish
         assert_eq!(
             block_manager
                 .read()