+++ CHANGELOG.md
@@ -28,6 +28,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - RS-262: Reduce number of open/close file operations, [PR-453](https://github.com/reductstore/reductstore/pull/453)
 
+### Fixed
+
+- Fix removing invalid blocks at start, [PR-454](https://github.com/reductstore/reductstore/pull/454)
+
 ## [1.9.5] - 2024-04-08
 
 ### Fixed